ComputeLD: Compute Linkage Disequilibrium (LD) for SNP Data

This function computes linkage disequilibrium (LD) statistics for SNP data using PLINK. It allows for computation across all SNPs or within specific chromosomes.

Usage

ComputeLD(DataDir, ResultDir, finput, ByCHR = FALSE, CHRnum = NULL, r2_LD)

Arguments

DataDir

Character string representing the file path of the input PLINK binary files.

ResultDir

Character string for the file path where all output files will be stored, defaulting to a temporary directory.

finput

Character string specifying the prefix of the input PLINK binary files.

ByCHR

Logical indicating whether to perform the computation by chromosome. The default is FALSE.

CHRnum

If ByCHR is TRUE, specifies the chromosome number for which LD should be computed. The default is NULL.

r2_LD

The threshold for r-squared LD values to report in the output.

Value

Returns a data frame containing the computed LD values among SNPs, read from the output file generated by PLINK.

Examples

# \donttest{
ComputeLD(
  DataDir = system.file("extdata", package = "GXwasR"), ResultDir = tempdir(),
  finput = "GXwasR_example", ByCHR = TRUE, CHRnum = 1, r2_LD = 0.2
)
#>     CHR_A     BP_A       SNP_A CHR_B     BP_B       SNP_B       R2
#> 1       1  1617443 rs112033089     1  1711414    rs867207 0.247273
#> 2       1  1821625   rs6603805     1  1875530   rs2803333 0.220796
#> 3       1  2793754 rs113572417     1  2806390 rs112163476 1.000000
#> 4       1  2838513  rs78083445     1  2845752  rs10797367 0.476771
#> 5       1  4520285   rs6426412     1  4549482    rs667500 0.228998
#> 6       1  4558972  rs13374152     1  4580135  rs13375764 0.249483
#> 7       1  5579571 rs113494976     1  5623159  rs61762104 0.227276
#> 8       1  5971455   rs6694744     1  6089881   rs1295102 0.242030
#> 9       1  6150893  rs10864358     1  6190611   rs1883764 0.212908
#> 10      1  6612933  rs12031598     1  6615690    rs925745 0.721959
#> 11      1  6612933  rs12031598     1  6631431  rs11587438 0.555581
#> 12      1  6615690    rs925745     1  6631431  rs11587438 0.628743
#> 13      1  7121898  rs11588586     1  7125002  rs11587771 0.295578
#> 14      1  7489495  rs58195072     1  7499159  rs11591156 0.622699
#> 15      1  7719653  rs17031274     1  7720431  rs72630555 0.991012
#> 16      1  8413839   rs1058790     1  8575064   rs6577492 0.327270
#> 17      1  8413839   rs1058790     1  8664922   rs1038008 0.622653
#> 18      1  8413839   rs1058790     1  8961530  rs12139087 0.347709
#> 19      1  8575064   rs6577492     1  8579301   rs7539382 0.419163
#> 20      1  8575064   rs6577492     1  8598005  rs10492966 0.419163
#> 21      1  8575064   rs6577492     1  8664922   rs1038008 0.337475
#> 22      1  8579301   rs7539382     1  8598005  rs10492966 1.000000
#> 23      1  8664922   rs1038008     1  8961530  rs12139087 0.444367
#> 24      1  9325751   rs9434745     1  9328808   rs3207316 0.668239
#> 25      1 10047519 rs150233209     1 10103131  rs11804193 0.327184
#> 26      1 10626922  rs11121594     1 10633543   rs6664134 1.000000
#> 27      1 11159823  rs12133922     1 11189191   rs3730379 0.933908
#> 28      1 11460626  rs72860274     1 11469860  rs75406294 1.000000
#> 29      1 11529251   rs2002891     1 11586025   rs2817635 0.208781
#> 30      1 11654009   rs2072998     1 11664772   rs7555966 0.453735
#> 31      1 12917418  rs56203441     1 12939159 rs143272699 0.332287
#> 32      1 13859410  rs75184403     1 13869073 rs116944585 0.956848
#> 33      1 14137801   rs2744690     1 14149249   rs2744682 0.785900
#> 34      1 14253320  rs16853159     1 14253821  rs10927959 0.323352
#> 35      1 14284508  rs12136489     1 14320465   rs6698763 0.741381
#> 36      1 14448388   rs4662106     1 14453475   rs7519955 0.567127
#> 37      1 15792426   rs2901964     1 15870118  rs72645808 0.520998
#> 38      1 15870118  rs72645808     1 16003404   rs2862154 0.285712
#> 39      1 16133525  rs56137831     1 16373861   rs7366864 0.213338
#> 40      1 17399432    rs761428     1 17402883   rs2647202 0.450326
#> 41      1 17782957   rs2489617     1 17832325   rs2253863 0.264336
#> 42      1 17832325   rs2253863     1 17848984   rs2247293 0.327294
#> 43      1 18652439  rs12725852     1 18660295    rs558986 0.226983
#> 44      1 19356651  rs72648090     1 19371099  rs61763740 0.496044
#> 45      1 19356651  rs72648090     1 19563900  rs16862632 0.201236
#> 46      1 19371099  rs61763740     1 19563900  rs16862632 0.330148
#> 47      1 19393387  rs10799666     1 19498443  rs12566296 0.264209
#> 48      1 19498443  rs12566296     1 19571875  rs12091991 0.539184
#> 49      1 20253293    rs818966     1 20255992   rs6660540 0.210364
#> 50      1 20253293    rs818966     1 20274703  rs61769120 0.707655
#> 51      1 20339742    rs818668     1 20415099  rs11573279 0.229352
#> 52      1 20615095   rs1022470     1 20648379   rs6426633 0.369271
#> 53      1 21170811  rs12042978     1 21286391   rs4654891 0.512019
#> 54      1 21170811  rs12042978     1 21339644   rs6681064 0.562912
#> 55      1 21286391   rs4654891     1 21339644   rs6681064 0.918894
#> 56      1 21744742  rs12240142     1 21745049   rs7417849 0.689853
#> 57      1 22159774   rs2229486     1 22173736  rs35921133 0.710236
#> 58      1 22211217   rs2290493     1 22290623  rs35806133 0.277234
#> 59      1 22211217   rs2290493     1 22320801  rs12404110 0.234427
#> 60      1 22290623  rs35806133     1 22299193  rs55953638 0.507805
#> 61      1 22290623  rs35806133     1 22311348  rs61777960 0.224208
#> 62      1 22290623  rs35806133     1 22320801  rs12404110 0.767104
#> 63      1 22290623  rs35806133     1 22342023   rs2865211 0.276409
#> 64      1 22299193  rs55953638     1 22311348  rs61777960 0.430641
#> 65      1 22299193  rs55953638     1 22320801  rs12404110 0.557490
#> 66      1 22299193  rs55953638     1 22342023   rs2865211 0.481588
#> 67      1 22309437  rs12128201     1 22312169 rs142271775 1.000000
#> 68      1 22311348  rs61777960     1 22320801  rs12404110 0.271613
#> 69      1 22311348  rs61777960     1 22342023   rs2865211 0.335726
#> 70      1 22320801  rs12404110     1 22342023   rs2865211 0.339822
#> 71      1 22918109  rs61767310     1 22923785  rs11582356 0.957708
#> 72      1 22918109  rs61767310     1 22924364  rs72651347 0.915751
#> 73      1 22923029    rs209689     1 22982406    rs291982 0.328811
#> 74      1 22923785  rs11582356     1 22924364  rs72651347 0.956190
#> 75      1 23709759  rs12131566     1 23733136  rs77282836 0.210021
#> 76      1 23729070   rs1615169     1 23733136  rs77282836 0.827068
#> 77      1 23973908   rs7527966     1 24199642   rs6663474 0.217306
#> 78      1 24036485  rs71640808     1 24079193  rs34032899 0.926481
#> 79      1 24062987  rs10917420     1 24199642   rs6663474 0.486401
#> 80      1 24745070    rs878751     1 24751197   rs4648983 0.952039
#> 81      1 24745070    rs878751     1 24754520   rs2223503 0.255806
#> 82      1 24745070    rs878751     1 24838493    rs196402 0.310936
#> 83      1 24751197   rs4648983     1 24754520   rs2223503 0.250860
#> 84      1 24751197   rs4648983     1 24838493    rs196402 0.317169
#> 85      1 25707462 rs144719257     1 26042492  rs56228815 1.000000
#> 86      1 26085398  rs12130495     1 26096132  rs12134148 0.735351
#> 87      1 26309937  rs12138372     1 26373756  rs12045468 0.434531
#> 88      1 26574330  rs12086573     1 26648034   rs4526602 0.624276
#> 89      1 28144171  rs57575116     1 28164029   rs6685626 1.000000
#> 90      1 28289538   rs6598915     1 28403989  rs11247749 0.241758
#> 91      1 28706949   rs4908348     1 28769814  rs35506828 0.756975
#> 92      1 28706949   rs4908348     1 28811547   rs6426348 0.941870
#> 93      1 28706949   rs4908348     1 28997257  rs11585196 0.295424
#> 94      1 28706949   rs4908348     1 29167823   rs2318774 0.355060
#> 95      1 28769814  rs35506828     1 28811547   rs6426348 0.816757
#> 96      1 28769814  rs35506828     1 28997257  rs11585196 0.257026
#> 97      1 28769814  rs35506828     1 29167823   rs2318774 0.264418
#> 98      1 28811547   rs6426348     1 28997257  rs11585196 0.319151
#> 99      1 28811547   rs6426348     1 29167823   rs2318774 0.347944
#> 100     1 28997257  rs11585196     1 29167823   rs2318774 0.256553
#> 101     1 29219755    rs157227     1 29448596 rs112665256 0.286020
#> 102     1 29219755    rs157227     1 29466867  rs35585334 0.691829
#> 103     1 29243978    rs183958     1 29408145   rs6694981 0.328897
#> 104     1 29448596 rs112665256     1 29466867  rs35585334 0.294173
#> 105     1 30806318  rs59867706     1 30871154 rs116632408 1.000000
#> 106     1 31035525 rs114580609     1 31078488  rs12088490 0.507858
#> 107     1 31035525 rs114580609     1 31135173  rs10914177 0.214722
#> 108     1 31078488  rs12088490     1 31135173  rs10914177 0.434815
#> 109     1 31771241  rs59195333     1 31834173  rs35068862 0.250578
#> 110     1 31771241  rs59195333     1 31859959 rs114737875 0.244093
#> 111     1 31834173  rs35068862     1 31859959 rs114737875 0.739747
#> 112     1 31919275   rs6425753     1 31949129  rs10914407 0.250094
#> 113     1 33626284    rs607008     1 33704214  rs66529020 0.535244
#> 114     1 33626284    rs607008     1 33724895   rs6670552 0.411954
#> 115     1 33704214  rs66529020     1 33724895   rs6670552 0.787914
#> 116     1 33765272   rs1132790     1 33889998  rs10914707 0.343292
#> 117     1 34963266   rs3845484     1 35028987  rs10914973 0.270021
#> 118     1 35119991   rs1415025     1 35137337  rs12045818 0.915699
#> 119     1 36620545  rs77650541     1 36649401  rs77014768 1.000000
#> 120     1 37631514  rs12141503     1 37677351 rs147107974 0.285151
#> 121     1 39279037  rs16825524     1 39285159  rs12130089 0.901441
#> 122     1 39628411  rs12117785     1 39803567   rs3121889 0.635493
#> 123     1 39638980 rs117211032     1 39868832  rs59544959 1.000000
#> 124     1 40519353  rs58483454     1 40576602  rs12756828 0.448901
#> 125     1 40786590 rs143168361     1 40786824   rs7513628 0.398605
#> 126     1 40978471  rs80234044     1 40981230   rs3795344 1.000000
#> 127     1 41064557    rs682981     1 41223414    rs928508 0.292761
#> 128     1 41190354   rs4660448     1 41223414    rs928508 0.430301
#> 129     1 41761080   rs7548367     1 41761737  rs12040690 0.496093
#> 130     1 41994331 rs115461457     1 41995339   rs2810550 0.966251
#> 131     1 41994331 rs115461457     1 42016774   rs9439045 0.247669
#> 132     1 41995339   rs2810550     1 42016774   rs9439045 0.256489
#> 133     1 42016774   rs9439045     1 42091517  rs10890148 0.222535
#> 134     1 42711449    rs343391     1 42788608   rs2479455 1.000000
#> 135     1 43182923  rs55751014     1 43219502  rs11802274 0.485192
#> 136     1 43794764   rs1199037     1 43901011   rs2039531 0.223120
#> 137     1 44089940    rs559575     1 44129779    rs669446 0.692893
#> 138     1 46248977   rs7514192     1 46287286   rs6662306 0.337892
#> 139     1 46248977   rs7514192     1 46356148  rs10890363 0.933070
#> 140     1 46287286   rs6662306     1 46348646  rs79380721 0.292308
#> 141     1 46287286   rs6662306     1 46356148  rs10890363 0.314183
#> 142     1 46966921  rs60830572     1 47009343  rs10890403 0.245323
#>  [ reached 'max' / getOption("max.print") -- omitted 919 rows ]
# }