DiffZeroOne: Assessing the Z-score for deviation from one and zero.

This function tests the null hypothesis that a measured statistics (example: genetic correlation, rg for a trait) < 1 using a 1-tailed test compared with a normal distribution (z = (1 − measure statistics)/Standard error). For multiple tests, users are encouraged to apply a Bonferroni multiple-testing correction.

DiffZeroOne(inputdata, diffzero = TRUE, diffone = TRUE)

Arguments

inputdata

A dataframe object, containing three columns:

Trait (i.e., the phenotype of interest)

Stat (i.e., the measured statistics)

SE (i.e., the standard error of the measured statistics)

diffzero

Boolean value, TRUE or FALSE, specifying to perform deviation from 0 test.

diffone

Boolean value, TRUE or FALSE, specifying to perform deviation from 1 test.

Value

A dataframe with columns:

Trait

Stat

SE

P0 (i.e, p-value for deviation from zero test)

P1 (i.e., p-value for deviation from 1 test)

Examples

data("Example_rgdata", package = "GXwasR")
inputdata <- Example_rgdata
colnames(inputdata) <- c("Trait", "Stat", "SE")
x <- DiffZeroOne(inputdata, FALSE, TRUE)