Download Hapmap phase 3 and 1000 Genome phase 3 Data

Downloads reference data sets from specified URLs based on the reference dataset name and working directory. Currently supports 'HapMapIII_NCBI36' and 'ThousandGenome'.

Usage

Download_reference(refdata, wdir = tempdir())

Arguments

refdata

A character string specifying the reference dataset to download. Should be one of 'HapMapIII_NCBI36' or 'ThousandGenome'.

wdir

A character string specifying the working directory where the reference data will be downloaded and extracted.

Value

Invisible. The function prints a message upon successful download and extraction of the reference data.

Examples

if (FALSE) { # \dontrun{
Download_reference("ThousandGenome", tempdir())
Download_reference("HapMapIII_NCBI36", tempdir())
} # }