MergeRegion: Merging two sets of plink binary files.

This function combines the two genotype datasets based on either common SNPs or all the SNPs between them.

Usage

MergeRegion(
  DataDir,
  ResultDir,
  finput1,
  finput2,
  foutput,
  use_common_snps = TRUE
)

Arguments

DataDir

A character string for the file path of the input plink binary files.

ResultDir

A character string for the file path where all output files will be stored. The default is tempdir().

finput1

Character string, specifying the prefix of the first input PLINK binary files.

finput2

Character string, specifying the prefix of the first input PLINK binary files.

foutput

Character string, specifying the prefix of the output PLINK binary files if filtering option for the SNPs is chosen. The default is "FALSE".

use_common_snps

Boolean value, TRUE or FALSE, specifying to use common SNPs for merging or to use all the SNPs.

Value

NULL

The output plink files will be saved in ResultDir.

Author

Banabithi Bose

Examples

# Not Run
# DataDir <- system.file("extdata", package = "GXwasR")
# ResultDir <- tempdir()
# finput1 <- "GXwasR_example"
# finput2 <- "GXwasR_example_imputed"
# foutput <- "Test_output"
# Not Run
# y <- MergeRegion(DataDir, ResultDir, finput1, finput2, foutput,  use_common_snps = TRUE)