plinkVCF: Converting VCF files to plink binary files and vice-versa.

This function performs the conversion between VCF files to plink binary formats.

For VCF to plink files conversion, if you do not specify any FAM file when you are converting from VCF to plink format, then plink will just create a 'dummy' FAM file with the same name as your dataset with missing phenotypes and missing sex.

Usage

plinkVCF(
  DataDir,
  ResultDir = tempdir(),
  finput,
  foutput,
  VtoP = FALSE,
  PtoV = TRUE,
  Famfile = NULL,
  PVbyCHR = TRUE
)

Arguments

DataDir

A character string for the file path of the input plink binary files and all other input files.

ResultDir

A character string for the file path where all output files will be stored. The default is tempdir().

finput

Character string, specifying the prefix of the input plink binary files or vcf files. This file needs to be in DataDir.

foutput

Character string, specifying the prefix of the output plink binary files if filtering option for the SNPs is chosen. The default is "FALSE".

VtoP

Boolean value, TRUE or FALSE, specifying the conversion of VCF files to plink binary files or not. The default is TRUE.

PtoV

Boolean value, TRUE or FALSE, specifying the conversion of plink binary files to VCF files or not. The default is TRUE.

Famfile

Character string, specifying the name of the original .fam file if VtoP was set to be TRUE. This file needs to be in DataDir. The default is NULL.

PVbyCHR

Boolean value, TRUE or FALSE specifying to do the plink to vcf conversion chromosome-wise or not. The default is TRUE.

Value

NULL

The output files will be saved in ResultDir.

Author

Banabithi Bose

Examples

finput <- "GXwasR_example" # Plink file
foutput <- "GXwasR_example1"
DataDir <- GXwasR:::GXwasR_data()
ResultDir <- tempdir()
PtoV <- TRUE
VtoP <- FALSE
Famfile <- NULL
PVbyCHR <- FALSE
plinkVCF(DataDir, ResultDir, finput, foutput, VtoP, PtoV, Famfile, PVbyCHR)
#> An error occurred: An error occurred while executing Plink: PLINK binary not found.
#> ✖ Attempted to locate the 'PLINK_PATH' environment variable and 'plink' in system PATH.
#> ℹ Ensure PLINK is available and executable (bioinformatics version).
#> ℹ You can permanently set the path to 'plink' by running:
#>   usethis::edit_r_environ()  # then add a line like: PLINK_PATH=/full/path/to/plink
#> NULL