SexDiffZscore: Z-score-based sex difference test.

This function calculates the difference in any kind of measured entities,(example: including SNP heritability estimate, genetic correlation, and GWAS β values) between sexes using a Z-score and its associated p-value statistic. When STAT/SE is normally distributed and the test statistics are independent in sex, the test is well calibrated. If the statistics are positively correlated, this test is conservative (1).

We could define SNPs with SDEs as those variants at the extreme ends of the distribution with an absolute value of the Z-score greater than 3(|Z-score| > 3), which is roughly equivalent to p <10−3, and represents 0.3% of all tested SNPs. The input dataframes should only include X-chromosome in order to obtain results for sex differences based solely on X-linked loci.

Usage

SexDiffZscore(inputdata)

Arguments

inputdata

A dataframe with five columns:

ID (i.e., SNP ID or the phenotype of interest, etc.)

Fstat (i.e., the measured statistics in females)

Fse (i.e., the standard error of the measured statistics in females)

Mstat (i.e., the measured statistics in males)

Mse (i.e., the standard error of the measured statistics in males)

Value

Original input dataframe with:

Zscore (i.e., Z-score),

p (i.e., p-value) and

adjP (i.e., Bonferroni corrected p-value) columns added.

Examples

data("GXwasRData")
inputdata <- Example_h2data
x <- SexDiffZscore(inputdata)